Use 64 bit ints for bytes used/reserved and ratio
This fixes an integer overflow in the free/used ratio calculation for
binlog compaction, and allows for tracking more than ~2.1GB of free or
used data in beanstalkd.

Fixes kr/beanstalkd#276

+++ dat.h
@@ -327,8 +327,8 @@ struct Wal {
     File   *tail;
     int    nfile;
     int    next;
-    int    resv;  // bytes reserved
-    int    alive; // bytes in use
+    int64  resv;  // bytes reserved
+    int64  alive; // bytes in use
     int64  nmig;  // migrations
     int64  nrec;  // records written ever
     int    wantsync;
